this commit will enable navigation with react router
