+++ frontend/src/BoardContext.js
@@ -1,3 +1,3 @@
-import {createContext, useContext} from 'react'
-export const BoardContext = createContext();
-export const useBoardContext = () => useContext(BoardContext)
+// import {createContext, useContext} from 'react'
+// export const BoardContext = createContext();
+// export const useBoardContext = () => useContext(BoardContext)
